Remove PHPLdapMyAdmin
The container that provides PHPLdapMyAdmin is ancient, no longer
maintained, and likely compromised.  Removal for security reasons
alone make sense, but also this tool is very old.  There is a newer
web GUI tool that is maintained and running in a more up to date
container in a separate PR.